Allow numeric positional parameters to be negative
The problem was a command line like this:

calibrate -45.333

That "-45.333" would be picked up as an option, when it was actually
intended to be a position parameter that is a negative number. This
is now allowed as a special case, but only if there is not an
option with that exact string value.